requirements.txt
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
Levenshtein==0.25.1
nltk==3.8.2
python-Levenshtein==0.25.1
python-telegram-bot==21.4
rapidfuzz==3.9.6
regex==2024.7.24
sniffio==1.3.1
tqdm==4.66.5
